If paket bootstrapper fails, suggest common fix
One of the most common reasons for the Paket bootstrapper to fail on
Linux is if the Mono installation is brand-new -- because out of the
box, Mono does not trust *any* X.509 root certificates. Therefore,
GitHub's SSL certificate won't be trusted and Paket won't be downloaded.

The fix most often recommended is to run:

    mozroots --import --sync

which imports the entire root SSL certificate store trusted by the
Mozilla project. This is over 100 certificates, so not everyone will
want to run this command. (And we certainly shouldn't be running it for
them.) But for someone who is new to F# and Mono, this suggestion might
save them hours of frustrated Google searches.

Note that we try to only print this message if it looks like it is
actually true for the user's situation, by using Mono's certificate
manager to count the number of trusted X.509 certificates. We compare to
1 rather than 0 because the word "X.509" will appear once in the header
of certmgr's output even on a system with no trusted root certs.